+++ src/JsonApiDotNetCore/Configuration/Validation/JsonApiMetaDataProvider.cs
@@ -1,17 +1,14 @@
-// Copyright (c) .NET Foundation. All rights reserved.
-// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
-
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.ModelBinding;
 using Microsoft.AspNetCore.Mvc.ModelBinding.Metadata;
 using Microsoft.Extensions.Options;
 
-namespace JsonApiDotNetCore.Configuration
+namespace JsonApiDotNetCore.Configuration.Validation
 {
     /// <summary>
     /// A default implementation of <see cref="IModelMetadataProvider"/> based on reflection.
     /// </summary>
-    public class JsonApiModelMetadataProvider : DefaultModelMetadataProvider
+    internal class JsonApiModelMetadataProvider : DefaultModelMetadataProvider
     {
         public JsonApiModelMetadataProvider(ICompositeMetadataDetailsProvider detailsProvider)
             : base(detailsProvider) { }
